+++ core/server/web/site/routes.js
@@ -1,9 +1,9 @@
 const debug = require('@tryghost/debug')('routing');
 const routing = require('../../../frontend/services/routing');
-const frontendSettings = require('../../../frontend/services/settings');
+const routeSettings = require('../../services/route-settings');
 
 module.exports = function siteRoutes(options = {}) {
     debug('site Routes', options);
-    options.routerSettings = frontendSettings.get();
+    options.routerSettings = routeSettings.loadRouteSettingsSync();
     return routing.bootstrap.init(options);
 };
