+++ test/regression/site/intergration-web-site/v2.test.js
@@ -7,7 +7,7 @@ const mockUtils = require('../../../utils/mocks');
 const configUtils = require('../../../utils/configUtils');
 const urlUtils = require('../../../utils/urlUtils');
 const appService = require('../../../../core/frontend/services/apps');
-const frontendSettingsService = require('../../../../core/frontend/services/settings');
+const routeSettingsService = require('../../../../core/server/services/route-settings');
 const themeEngine = require('../../../../core/frontend/services/theme-engine');
 const siteApp = require('../../../../core/server/web/parent/app');
 
@@ -368,7 +368,7 @@ describe('Integration - Web - Site v2', function () {
     describe('extended routes.yaml: collections', function () {
         describe('2 collections', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {
                         '/': {templates: ['home']}
                     },
@@ -497,7 +497,7 @@ describe('Integration - Web - Site v2', function () {
 
         describe('no collections', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {
                         '/something/': {templates: ['something']}
                     },
@@ -546,7 +546,7 @@ describe('Integration - Web - Site v2', function () {
 
         describe('static permalink route', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {},
 
                     collections: {
@@ -650,7 +650,7 @@ describe('Integration - Web - Site v2', function () {
 
         describe('primary author permalink', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {},
 
                     collections: {
@@ -733,7 +733,7 @@ describe('Integration - Web - Site v2', function () {
 
         describe('primary tag permalink', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {},
 
                     collections: {
@@ -831,7 +831,7 @@ describe('Integration - Web - Site v2', function () {
 
         describe('collection/routes with data key', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {
                         '/my-page/': {
                             data: {
@@ -1000,7 +1000,7 @@ describe('Integration - Web - Site v2', function () {
     describe('extended routes.yaml: templates', function () {
         describe('default template, no template', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {},
 
                     collections: {
@@ -1070,7 +1070,7 @@ describe('Integration - Web - Site v2', function () {
 
         describe('two templates', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {},
 
                     collections: {
@@ -1122,7 +1122,7 @@ describe('Integration - Web - Site v2', function () {
 
         describe('home.hbs priority', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {},
 
                     collections: {
@@ -1201,7 +1201,7 @@ describe('Integration - Web - Site v2', function () {
             before(function () {
                 testUtils.integrationTesting.defaultMocks(sinon, {theme: 'test-theme-channels'});
 
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {
                         '/channel1/': {
                             controller: 'channel',
@@ -1535,7 +1535,7 @@ describe('Integration - Web - Site v2', function () {
 
     describe('extended routes.yaml (5): rss override', function () {
         before(function () {
-            sinon.stub(frontendSettingsService, 'get').returns({
+            sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                 routes: {
                     '/podcast/rss/': {
                         templates: ['podcast/rss'],
