+++ test/regression/site/intergration-web-site/v3.test.js
@@ -7,7 +7,7 @@ const mockUtils = require('../../../utils/mocks');
 const configUtils = require('../../../utils/configUtils');
 const urlUtils = require('../../../utils/urlUtils');
 const appService = require('../../../../core/frontend/services/apps');
-const frontendSettingsService = require('../../../../core/frontend/services/settings');
+const routeSettingsService = require('../../../../core/server/services/route-settings');
 const themeEngine = require('../../../../core/frontend/services/theme-engine');
 const siteApp = require('../../../../core/server/web/parent/app');
 
@@ -368,7 +368,7 @@ describe('Integration - Web - Site v3', function () {
     describe('extended routes.yaml: collections', function () {
         describe('2 collections', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {
                         '/': {
                             templates: ['home']
@@ -499,7 +499,7 @@ describe('Integration - Web - Site v3', function () {
 
         describe('no collections', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {
                         '/something/': {templates: ['something']}
                     },
@@ -548,7 +548,7 @@ describe('Integration - Web - Site v3', function () {
 
         describe('static permalink route', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {},
 
                     collections: {
@@ -652,7 +652,7 @@ describe('Integration - Web - Site v3', function () {
 
         describe('primary author permalink', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {},
 
                     collections: {
@@ -735,7 +735,7 @@ describe('Integration - Web - Site v3', function () {
 
         describe('primary tag permalink', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {},
 
                     collections: {
@@ -833,7 +833,7 @@ describe('Integration - Web - Site v3', function () {
 
         describe('collection/routes with data key', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {
                         '/my-page/': {
                             data: {
@@ -1002,7 +1002,7 @@ describe('Integration - Web - Site v3', function () {
     describe('extended routes.yaml: templates', function () {
         describe('default template, no template', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {},
 
                     collections: {
@@ -1072,7 +1072,7 @@ describe('Integration - Web - Site v3', function () {
 
         describe('two templates', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {},
 
                     collections: {
@@ -1124,7 +1124,7 @@ describe('Integration - Web - Site v3', function () {
 
         describe('home.hbs priority', function () {
             before(function () {
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {},
 
                     collections: {
@@ -1203,7 +1203,7 @@ describe('Integration - Web - Site v3', function () {
             before(function () {
                 testUtils.integrationTesting.defaultMocks(sinon, {theme: 'test-theme-channels'});
 
-                sinon.stub(frontendSettingsService, 'get').returns({
+                sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                     routes: {
                         '/channel1/': {
                             controller: 'channel',
@@ -1537,7 +1537,7 @@ describe('Integration - Web - Site v3', function () {
 
     describe('extended routes.yaml (5): rss override', function () {
         before(function () {
-            sinon.stub(frontendSettingsService, 'get').returns({
+            sinon.stub(routeSettingsService, 'loadRouteSettingsSync').returns({
                 routes: {
                     '/about/': 'about',
                     '/podcast/rss/': {
